+++ python/hooks/check.bash
@@ -1,4 +1,8 @@
 #! /usr/bin/env bash
+# Shellcheck has some false positives
+# seems to be multiple issues on github
+# so ignore for now. Please retest later.
+# shellcheck disable=SC2030,SC2031
 
 printStatus() {
     case "$1" in
@@ -14,64 +18,141 @@ printStatus() {
     esac
 }
 
+runTool() {
+    local tool
+    tool="$1"
+    shift 1
+
+    local displayTool
+    displayTool="$1"
+    shift 1
+
+    local disableVar
+    disableVar="dontRun${tool^}"
+    local statusVar
+    statusVar="${tool}Status"
+
+    if [ -z "${!disableVar:-}" ] && [[ "$(command -v "$tool")" =~ ^/nix/store/.*$ ]]; then
+        echo -e "\n\x1b[1;36m${displayTool^}:\x1b[0m"
+        "$tool" "$@" 2>&1 | sed 's/^/  /'
+        declare -xg "$statusVar"=$?
+    else
+        echo "$tool is disabled."
+    fi
+}
+
 standardTests() (
     # clean up after pip
     rm -rf build/
 
     set +e
-    echo -e "\n\x1b[1;36mBlack:\x1b[0m"
-    if [[ "$(command -v black)" =~ ^/nix/store/.*$ ]]; then
-        # shellcheck disable=SC2086
-        black ${blackArgs:-} --check .  2>&1 | sed 's/^/  /'
-        blackStatus=$?
-    else
-        echo "  Black not supported on platform ${system:-unknown}."
-    fi
+    set -o pipefail
 
-    echo -e "\n\x1b[1;36mIsort:\x1b[0m"
     # shellcheck disable=SC2086
-    isort ${isortArgs:-} --check . 2>&1 | sed 's/^/  /'
-    isortStatus=$?
+    runTool black black ${blackArgs:-} --check .
 
-    echo -e "\n\x1b[1;36mPylint:\x1b[0m"
+    # shellcheck disable=SC2086
+    runTool isort isort ${isortArgs:-} --check .
+
     # shellcheck disable=SC2046,SC2086
-    HOME=$TMP pylint ${pylintArgs:-} --recursive=y . 2>&1 | sed 's/^/  /'
-    pylintStatus=$?
+    HOME=$TMP runTool pylint pylint ${pylintArgs:-} --recursive=y .
 
-    echo -e "\n\x1b[1;36mFlake8:\x1b[0m"
     # shellcheck disable=SC2086
-    flake8 ${flake8Args:-} . 2>&1 | sed 's/^/  /'
-    flake8Status=$?
+    runTool flake8 flake8 ${flake8Args:-} .
 
-    echo -e "\n\x1b[1;36mMypy:\x1b[0m"
     # shellcheck disable=SC2086
-    mypy ${mypyArgs:-} . 2>&1 | sed 's/^/  /'
-    mypyStatus=$?
+    runTool mypy mypy ${mypyArgs:-} .
 
-    echo -e "\n\x1b[1;36mPytest:\x1b[0m"
     # shellcheck disable=SC2086
-    pytest ${pytestArgs:-} . 2>&1 | sed 's/^/  /'
-    pytestStatus=$?
+    runTool pytest pytest ${pytestArgs:-} .
 
     # no tests ran
-    if [ $pytestStatus -eq 5 ]; then
+    if [ "${pytestStatus:-0}" -eq 5 ]; then
+        local pytestStatus
         pytestStatus=0
     fi
 
     echo -e "Summary:
   black: $(printStatus "${blackStatus:-skipped}")
-  isort: $(printStatus $isortStatus)
-  pylint: $(printStatus $pylintStatus)
-  flake8: $(printStatus $flake8Status)
-  mypy: $(printStatus $mypyStatus)
-  pytest: $(printStatus $pytestStatus)"
+  isort: $(printStatus "${isortStatus:-skipped}")
+  pylint: $(printStatus "${pylintStatus:-skipped}")
+  flake8: $(printStatus "${flake8Status:-skipped}")
+  mypy: $(printStatus "${mypyStatus:-skipped}")
+  pytest: $(printStatus "${pytestStatus:-skipped}")"
 
-    blackStatus=${blackStatus:-0}
+    : "${blackStatus:=0}" "${isortStatus:=0}" "${pylintStatus:=0}"
+    : "${flake8Status:=0}" "${mypyStatus:=0}" "${pytestStatus:=0}"
     exit $((blackStatus + isortStatus + pylintStatus + flake8Status + mypyStatus + pytestStatus))
+)
+
+ruffStandardTests() (
+    # clean up after pip
+    rm -rf build/
+
+    set +e
+    set -o pipefail
+
+    # shellcheck disable=SC2086
+    runTool ruff "Ruff Check" check ${ruffArgs:-} .
+    local ruffCheckStatus
+    ruffCheckStatus=${ruffStatus:-}
+
+    # shellcheck disable=SC2086
+    runTool ruff "Ruff Check Imports" check ${ruffArgs:-} --select I --diff .
+    local ruffImportsStatus
+    ruffImportsStatus=${ruffStatus:-}
+
+    # shellcheck disable=SC2086
+    runTool ruff "Ruff Format" format --diff ${ruffArgs:-} .
+    local ruffFormatStatus
+    ruffFormatStatus=${ruffStatus:-}
+
+    # shellcheck disable=SC2086
+    runTool mypy mypy ${mypyArgs:-} .
+
+    # shellcheck disable=SC2086
+    runTool pytest pytest ${pytestArgs:-} .
+
+    # no tests ran
+    if [ "${pytestStatus:-0}" -eq 5 ]; then
+        local pytestStatus
+        pytestStatus=0
+    fi
+
+    echo -e "Summary:
+  ruff check: $(printStatus "${ruffCheckStatus:-skipped}")
+  ruff imports: $(printStatus "${ruffImportsStatus:-skipped}")
+  ruff format: $(printStatus "${ruffFormatStatus:-skipped}")
+  mypy: $(printStatus "${mypyStatus:-skipped}")
+  pytest: $(printStatus "${pytestStatus:-skipped}")"
+
+    : "${ruffCheckStatus:=0}" "${ruffFormatStatus:=0}"
+    : "${mypyStatus:=0}" "${pytestStatus:=0}" "${ruffImportsStatus:=0}"
+    exit $((ruffCheckStatus + ruffFormatStatus + mypyStatus + pytestStatus + ruffImportsStatus))
 )
 
 # If there is a checkPhase declared, mk-python-component in nixpkgs will put it in
 # installCheckPhase so we use that phase as well (since this is executed later).
 if [ -n "${doStandardTests-}" ] && [ -z "${installCheckPhase-}" ]; then
-    installCheckPhase=standardTests
+    installCheckPhase=@defaultCheckPhase@
 fi
+
+runStandardFormat() {
+    black .
+    isort .
+}
+
+runRuffFormat() {
+    ruff format .
+    ruff check --select I --fix .
+}
+
+runFormat() {
+    if [ "${formatter:-}" = "standard" ]; then
+        runStandardFormat
+    elif [ "${formatter:-}" = "ruff" ] || [[ "${installCheckPhase:-}" =~ "ruffStandardTests" ]]; then
+        runRuffFormat
+    else
+        runStandardFormat
+    fi
+}